velican2/requirements.txt
psycopg2
django
django-allauth
django-admin-tools
django-markdownx
martor
markdown
pelican
Pillow
boto3
pyYAML
toml